fixup! Add validation to relevant models
