+++ app/models/user.rb
@@ -62,7 +62,7 @@ class User < ApplicationRecord
     dependent: :destroy
 
   include Token
-  include EmailBlocklistHelper
+  include EmailBlocklistValidation
 
   # As of Rails 8.0, `has_secure_password` generates a `password_reset_token`
   # method that shadows the explicit `password_reset_token` attribute.
@@ -99,8 +99,6 @@ class User < ApplicationRecord
     length: {maximum: 100},
     format: {with: /\A[^@ ]+@[^@ ]+\.[^@ ]+\Z/},
     uniqueness: {case_sensitive: false}
-
-  validate :email_blocklist
 
   validates :homepage,
     format: {
@@ -629,11 +627,5 @@ class User < ApplicationRecord
       .where("(votes.comment_id is not null and comments.user_id <> votes.user_id) OR " \
                  "(votes.comment_id is null and stories.user_id <> votes.user_id)")
       .order(id: :desc)
-  end
-
-  def email_blocklist
-    if email_on_blocklist?(email)
-      errors.add(:email, "disposable emails are blocked")
-    end
   end
 end
